Phase 3: implement Compute Runtime with CPU and Vulkan GPU backends
Compute Task:
- VectorAddTask (C[i]=A[i]+B[i], int32) - identical math on CPU and GPU,
  bit-exact comparison possible; Status enum + error objects (no exceptions),
  documented invalid-input policy (size mismatch and empty input rejected)

Runtime:
- Lifecycle: initialize() -> execute() -> shutdown(); non-copyable (RAII)
- Explicit backend selection ('cpu', 'vulkan'); no auto-scheduling (Phase 7)
- GPU unavailability is a normal environment state, never a crash: reason
  exposed via backend_unavailable_reason(), CPU path keeps working

CPU backend:
- Always-available reference implementation; result is the verification
  baseline for GPU execution

Vulkan GPU backend (real compute pipeline, no graphics/WSI):
- instance -> physical device selection (prefers real GPUs, honestly types
  software implementations) -> logical device + compute queue -> storage
  buffers -> upload -> dispatch (embedded SPIR-V) -> readback -> verification
- RAII everywhere: PIMPL Vulkan state, per-execution ScopedBuffer pairs,
  idempotent destroy on every failure path and at shutdown
- SPIR-V pre-compiled from shaders/vector_add.comp and embedded
  (scripts/embed_spv.py) - no shader compiler needed at build time,
  nothing downloaded at runtime
- Compiles as an explicit stub when Vulkan is not found: CPU-only builds
  work without the Vulkan SDK (VORTYX_ENABLE_VULKAN option, auto-detected)

Device abstraction extensions:
- DeviceType::SoftwareGpu so software Vulkan devices are never labeled as
  real GPUs; Mesa (0x10005) vendor mapping; 32-bit-safe vendor id handling

Tests (CPU/GPU paths separated):
- ComputeCpuTest: lifecycle, correctness at sizes 4/16/1024/10007, invalid
  inputs, unknown/unavailable backends, shutdown/re-init - passes everywhere
- ComputeGpuTest: real GPU execution when a Vulkan device exists (incl.
  non-workgroup-multiple sizes, determinism, re-init); explicit SKIP note
  otherwise - never a fake success

Verified locally with Mesa lavapipe software Vulkan device: GPU vector
addition matches CPU reference exactly. Version 0.3.0 everywhere.

+++ src/core/device/platform_linux.cpp
@@ -172,7 +172,7 @@ DiscoveryResult discover_gpus_platform() {
         gpu.type = DeviceType::Gpu;
         gpu.backend = "linux-sysfs";
 
-        const std::string vendor = pci_vendor_name(static_cast<std::uint16_t>(vendor_id));
+        const std::string vendor = pci_vendor_name(static_cast<std::uint32_t>(vendor_id));
         if (!vendor.empty()) {
             gpu.vendor = vendor;
         } else {
